cr-classifier/requirements.txt
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.3
opencv-python==4.2.0.34
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
psycopg2-binary=2.8.5